Started moving methods to object level.
